+++ client_src/js/octoprint/octoprint-plugin-commands.js
@@ -1,7 +1,7 @@
 import OctoPrintClient from "../lib/octoprint.js";
 import OctoFarmClient from "../services/octofarm-client.service.js";
 import UI from "../lib/functions/ui";
-import bulkActionsStates from "../printer-manager/bulk-actions.constants";
+import bulkActionsStates from "../pages/printer-manager/bulk-actions.constants";
 
 const printerBase = "printers";
 const printerInfoURL = "/printerInfo";