+++ client_src/js/system.runner.js
@@ -1,13 +1,13 @@
 import UI from "./lib/functions/ui.js";
 import Script from "./services/octofarm-scripts.service";
-import ClientSettings from "./system/client-settings";
+import ClientSettings from "./pages/system/client-settings";
 import {
   localStorageKeys,
   serverActionsElements,
   serverDatabaseKeys,
   userActionElements,
   returnSaveBtn
-} from "./system/server.options";
+} from "./pages/system/server.options";
 import {
   generateLogDumpFile,
   exportDatabases,
@@ -27,8 +27,8 @@ import {
   resetUserPassword,
   fillInEditInformation,
   setupOPFilamentManagerPluginSettings
-} from "./system/server.actions";
-import { serverBootBoxOptions } from "./system/utils/bootbox.options";
+} from "./pages/system/server.actions";
+import { serverBootBoxOptions } from "./pages/system/utils/bootbox.options";
 import { removeLocalStorage } from "./services/local-storage.service";
 
 // Setup Page